Rename `flurs.data.entity` to `flurs.types`
