+++ examples/faust/recommender.py
@@ -6,7 +6,7 @@ Faust Consumer
 Consume events and incrementally train a FluRS recommender.
 """
 
-from flurs.data.entity import User, Item, Event
+from flurs.types import User, Item, Event
 from flurs.recommender import MFRecommender
 import json
 import numpy as np
