+++ flurs/baseline/tests/test_popular.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 import numpy as np
 from numpy.testing import assert_array_equal
 
-from flurs.data.entity import User, Item, Event
+from flurs.types import User, Item, Event
 from flurs.baseline import Popular
