+++ flurs/types/__init__.py
@@ -1,0 +1,5 @@
+from .actor import User, Item
+from .event import Event
+
+
+__all__ = ["User", "Item", "Event"]
